dashboard/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
mysql-connector-python>=8.0
